This PR makes `nix build` complete in approximately half the time, by disabling unused outputs.

This PR also adds specialized targets for a `-O0` build and a type-checking-only build.

## Benchmarks

* The timings in the code are from a single run of measurements
* I also took repeated measurements with the command

  ```bash
  hyperfine -p "echo >> src/full/Agda/TypeChecking/Monad/Base.hs" "nix build"
  ```

  for the first few optimizations:

  | Edits to `flake.nix`  |         Time         |
  | --------------------: | -------------------- |
  | L37-L41 commented out | 604.551 s ±  7.446 s |
  | L38-L41 commented out | 345.304 s ± 12.439 s |
  | L39-L41 commented out | 266.819 s ± 10.924 s |
  | L40-L41 commented out | 273.623 s ±  6.797 s |
  |     L41 commented out | 272.290 s ±  4.692 s |

  I took these benchmarks on my development laptop, in order, with no other builds running.

  I didn't benchmark all subsets of optimizations, or all the targets, because it was taking too long.